change in Makefile and CMakeLists : pack ready-to-use targz for Windows and Linux (32 or 64)
git-svn-id: https://alm.lip6.fr/svn/agrum/trunk@219 fa2884e2-1fde-468d-ae8a-4cc2b5279982

+++ src/testunits/MultiDimAggregatorTestSuite.h
@@ -169,9 +169,9 @@ class MultiDimAggregratorsTestSuite: public CxxTest::TestSuite {
       //it is not allowed to change a value but can only be detected at the next access
       i.setFirst();
 
-      TS_ASSERT_THROWS( p.set( i, 3.0 ), gum::OperationNotAllowed );
+      TS_ASSERT_THROWS( p.set( i, 3 ), gum::OperationNotAllowed );
 
-      TS_ASSERT_THROWS( p.fill(( double )0 ), gum::OperationNotAllowed );
+      TS_ASSERT_THROWS( p.fill(0 ), gum::OperationNotAllowed );
     }
 
     void testBNwithMin() {
